Changes to not break the compilation on old systems.
