+++ src/data_provider/src/packages/packageLinuxDataRetriever.h
@@ -13,12 +13,10 @@
 #define _PACKAGE_LINUX_DATA_RETRIEVER_H
 
 #include <memory>
+#include "filesystemHelper.h"
 #include "json.hpp"
 #include "sharedDefs.h"
-#include "filesystemHelper.h"
 #include "utilsWrapperLinux.hpp"
-#include "packagesPYPI.hpp"
-#include "packagesNPM.hpp"
 
 /**
  * @brief Fills a JSON object with all available pacman-related information
@@ -94,9 +92,6 @@ class FactoryPackagesCreator<LinuxType::STANDARD> final
             {
                 getApkInfo(APK_DB_PATH, callback);
             }
-
-            PYPI().getPackages(UNIX_PYPI_DEFAULT_BASE_DIRS, callback);
-            NPM().getPackages(UNIX_NPM_DEFAULT_BASE_DIRS, callback);
         }
 };
