+++ src/data_provider/src/sysInfoLinux.cpp
@@ -12,6 +12,9 @@
 #include <iostream>
 #include <regex>
 #include <sys/utsname.h>
+#include "packages/packagesNPM.hpp"
+#include "packages/packagesPYPI.hpp"
+#include "packages/modernPackageDataRetriever.hpp"
 #include "sharedDefs.h"
 #include "stringHelper.h"
 #include "filesystemHelper.h"
@@ -594,6 +597,8 @@ void SysInfo::getProcessesInfo(std::function<void(nlohmann::json&)> callback) co
 void SysInfo::getPackages(std::function<void(nlohmann::json&)> callback) const
 {
     FactoryPackagesCreator<LINUX_TYPE>::getPackages(callback);
+    ModernFactoryPackagesCreator<PYPI<>, STANDARD_TYPE>::getPackages(UNIX_PYPI_DEFAULT_BASE_DIRS, callback);
+    ModernFactoryPackagesCreator<NPM<>, STANDARD_TYPE>::getPackages(UNIX_NPM_DEFAULT_BASE_DIRS, callback);
 }
 
 nlohmann::json SysInfo::getHotfixes() const
