+++ src/shared_modules/utils/tests/CMakeLists.txt
@@ -11,7 +11,15 @@ endif(COVERITY)
 # Enable testing
 enable_testing()
 
-set(CMAKE_CXX_STANDARD 17)
+# If is Linux, Windows or macOS > 10.15, use C++17
+if(CMAKE_SYSTEM_NAME STREQUAL "Linux" OR CMAKE_SYSTEM_NAME STREQUAL "Windows" OR (CMAKE_SYSTEM_NAME STREQUAL "Darwin" AND NOT CMAKE_HOST_SYSTEM_VERSION VERSION_LESS "19.0.0"))
+  set(CMAKE_CXX_STANDARD 17)
+  file(GLOB UTIL_CXX_UNITTEST_CPP17_SRC
+      "stdFileSystemHelper_test.cpp")
+else()
+  set(CMAKE_CXX_STANDARD 14)
+endif()
+
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(CMAKE_CXX_FLAGS "-Wall -Wextra")
@@ -66,6 +74,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
     add_executable(utils_unit_test
       ${UTIL_CXX_UNITTEST_COMMON_SRC}
       ${UTIL_CXX_UNITTEST_WINDOWS_SRC}
+      ${UTIL_CXX_UNITTEST_CPP17_SRC}
     )
     add_definitions(-DWIN32=1
                     -D_WIN32_WINNT=0x600)
@@ -90,11 +99,13 @@ else()
     if (APPLE)
       add_executable(utils_unit_test
         ${UTIL_CXX_UNITTEST_COMMON_SRC}
+        ${UTIL_CXX_UNITTEST_CPP17_SRC}
       )
     else()
       add_executable(utils_unit_test
         ${UTIL_CXX_UNITTEST_COMMON_SRC}
         ${UTIL_CXX_UNITTEST_LINUX_SRC}
+        ${UTIL_CXX_UNITTEST_CPP17_SRC}
       )
     endif(APPLE)
