+++ src/shared_modules/utils/tests/filesystemHelper_test.cpp
@@ -60,97 +60,5 @@ TEST_F(FilesystemUtilsTest, getFileBinaryContent)
     EXPECT_FALSE(binContent.empty());
 }
 
-TEST_F(FilesystemUtilsTest, FilesystemExpandSimpleWildcard)
-{
-    constexpr auto PATH_MATCH_SIZE { 2ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/dum*", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/dummy" || item == "/tmp/wazuh_test/dummy.txt");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-
-TEST_F(FilesystemUtilsTest, FilesystemExpandWildcard)
-{
-    constexpr auto PATH_MATCH_SIZE { 4ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/prefix_*_data/*", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/prefix_1_data/prefix1_1" ||
-                    item == "/tmp/wazuh_test/prefix_1_data/prefix1_2" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_1" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_2");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-
-TEST_F(FilesystemUtilsTest, FilesystemExpandWildcardWithPrefix)
-{
-    constexpr auto PATH_MATCH_SIZE { 4ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/prefix_*_data/prefix*", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/prefix_1_data/prefix1_1" ||
-                    item == "/tmp/wazuh_test/prefix_1_data/prefix1_2" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_1" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_2");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-
-TEST_F(FilesystemUtilsTest, FilesystemExpandWildcardWithSuffix)
-{
-    constexpr auto PATH_MATCH_SIZE { 2ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/prefix_*_data/*_1", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/prefix_1_data/prefix1_1" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_1");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-TEST_F(FilesystemUtilsTest, FilesystemExpandWildcardWithQuestionMark)
-{
-    constexpr auto PATH_MATCH_SIZE { 2ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/prefix_?_data/*_1", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/prefix_1_data/prefix1_1" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_1");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-
-TEST_F(FilesystemUtilsTest, FilesystemExpandWildcardWithQuestionMark2)
-{
-    constexpr auto PATH_MATCH_SIZE { 2ull };
-    std::vector<std::string> output;
-    Utils::expandAbsolutePath("/tmp/wazuh_test/prefix_*_data/prefix?*1", output);
-
-    for (const auto& item : output)
-    {
-        EXPECT_TRUE(item == "/tmp/wazuh_test/prefix_1_data/prefix1_1" ||
-                    item == "/tmp/wazuh_test/prefix_2_data/prefix2_1");
-    }
-
-    EXPECT_EQ(output.size(), PATH_MATCH_SIZE);
-}
-
 
 #endif
