+++ src/shared_modules/utils/tests/filesystemHelper_test.h
@@ -24,24 +24,5 @@ class FilesystemUtilsTest : public ::testing::Test
 
         void SetUp() override;
         void TearDown() override;
-
-        static void SetUpTestSuite()
-        {
-            std::filesystem::remove_all("/tmp/wazuh_test");
-            std::filesystem::create_directory("/tmp/wazuh_test");
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_1_data");
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_2_data");
-            std::filesystem::create_directory("/tmp/wazuh_test/dummy");
-            // Create dummy file
-            std::ofstream dummyFile("/tmp/wazuh_test/dummy.txt");
-            dummyFile << "dummy";
-            dummyFile.close();
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_1_data/prefix1_1");
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_1_data/prefix1_2");
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_2_data/prefix2_1");
-            std::filesystem::create_directory("/tmp/wazuh_test/prefix_2_data/prefix2_2");
-
-            std::this_thread::sleep_for(std::chrono::seconds(2));
-        }
 };
 #endif //FILESYSTEM_HELPER_TESTS_H
